added read me and finished crossing off spec.md files

+++ db/migrate/20180410182931_add_donations_to_supplies.rb
@@ -1,5 +1,5 @@
 class AddDonationsToSupplies < ActiveRecord::Migration
   def change
-    add_column :supplies, :donations, :integer, default: 0
+    add_column :supplies, :donations, :float, default: 0
   end
 end
